+++ History.md
@@ -5,6 +5,7 @@ Dalli Changelog
 =======
 
 - Fix all dalli\_store operations to handle nil options [#190]
+- Increment and decrement with :initial => nil now return nil (lawrencepit, #112)
 
 2.0.1
 =======